Change of storage location of templates and views
Now templates and views should be storage in directory of the same name
as app.
Django check the first directory if he found name matches or if we have
another template with the same name in diffrent app Django would be
unable to distinguish between them. That's why we should point Django
the right direcotry.