Add some fundamental functions
Login and logout with login reqiuerd now it is no nessesery to reset all
to logout only need to write in url /logout .

Change welcome_page now it is other function and view so it easy to
change if we needed and it more clear.

And of course `login_requierd` now works preety weel but I thing in
feathure we must change url to redirect.