Find_recipe_base_on_products version 2.0
Currently function return dict which key is recipe and value is list of
product what we needed. In next we should try to sort it by lenght of
value. That solutin give us posibility to show user first recipes that
he can make without buy new product and leater recipes with missing one
or more products.